refactor #186 linkuCreate Service 遺꾨━
