+++ src/main/java/com/umc/linkyou/service/Linku/LinkuServiceImpl.java
@@ -87,154 +87,6 @@ public class LinkuServiceImpl implements LinkuService {
 
     @Override
     @Transactional
-    public LinkuResponseDTO.LinkuCreateResult createLinku(Long userId, LinkuRequestDTO.LinkuCreateDTO dto, MultipartFile image) {
-        // 1) URL 정규화 & 검증 (비디오 링크 여부, URL 유효성 체크)
-        String normalizedLink = validateAndNormalizeUrl(dto.getLinku());
-
-        String domainTail = UrlValidUtils.extractDomainTail(normalizedLink);
-
-        // 2) AI 분류 실행 → Category + AI 키워드 반환
-        AiCategoryInfo aiInfo = resolveCategoryAndKeywords(normalizedLink);
-        Category category = aiInfo.category;
-        String aiKeywords = aiInfo.aiKeywords;
-
-        // 3) 감정(Emotion) 조회 (없으면 기본값)
-        Emotion emotion = resolveEmotion(dto.getEmotionId());
-
-        // 4) Domain 조회 (없으면 기본값)
-        Domain domain = resolveDomain(domainTail);
-
-        // 5) Linku 조회 또는 신규 생성
-        Linku linku = findOrCreateLinku(normalizedLink, category, domain, domainTail);
-
-        // 6) AI Article 존재 여부 확인하고 필요시 생성
-        createAiArticleIfNeeded(linku, category, emotion, aiKeywords);
-
-        // 7) 요청 보낸 사용자 조회
-        Users user = findUser(userId);
-
-        // 8) 이미지 저장 (파일 업로드 or 링크 이미지 추출)
-        String imageUrl = processImage(image, linku);
-
-        // 9) UsersLinku 생성 & 저장
-        UsersLinku usersLinku = createUsersLinku(user, linku, emotion, dto.getMemo(), imageUrl);
-
-        // 9-1) 링크 생성 후 "최근 열람 링크"에도 기록 추가
-        updateRecentViewedLinku(userId, linku.getLinkuId());
-
-        // 10) 폴더 조회 또는 신규 생성
-        Folder folder = findOrCreateFolder(userId, category);
-
-        // 11) LinkuFolder 생성 & 저장
-        LinkuFolder linkuFolder = LinkuConverter.toLinkuFolder(folder, usersLinku);
-        linkuFolderRepository.save(linkuFolder);
-
-        // 12) 응답 DTO 변환
-        LinkuResponseDTO.LinkuResultDTO resultDto =
-                LinkuConverter.toLinkuResultDTO(userId, linku, usersLinku, linkuFolder, category, domain, null, aiKeywords, null);
-
-        // 13) 최종 결과 반환
-        return LinkuResponseDTO.LinkuCreateResult.builder()
-                .data(resultDto)
-                .validUrl(UrlValidUtils.isURLConnectionOk(normalizedLink)) // URL 연결 가능 여부
-                .build();
-    }// 링큐 생성
-
-    //링큐 생성 편의 메소드 시작
-    // 1. URL 정규화 및 기본 검증(영상 여부, 유효성)
-    private String validateAndNormalizeUrl(String url) {
-        String normalized = UrlUtils.normalizeUrl(url);
-        if (UrlValidUtils.isVideoLink(normalized)) throw new GeneralException(ErrorStatus._LINKU_VIDEO_NOT_ALLOWED);
-        if (!UrlValidUtils.isValidUrl(url)) throw new GeneralException(ErrorStatus._LINKU_INVALID_URL);
-        return normalized;
-    }
-
-    // 2. AI 카테고리 분류 + Category 조회 결과 묶어서 반환
-    private record AiCategoryInfo(Category category, String aiKeywords) {}
-
-    private AiCategoryInfo resolveCategoryAndKeywords(String normalizedLink) {
-        OpenAICategoryClassifier.CategoryResult aiResult =
-                openAiCategoryClassifier.classifyCategoryByUrl(normalizedLink, categoryRepository.findAll());
-        Long aiCategoryId = (aiResult != null) ? aiResult.getCategoryId() : null;
-        String aiKeywords = (aiResult != null) ? aiResult.getKeywords() : null;
-        Category category = Optional.ofNullable(aiCategoryId)
-                .flatMap(categoryRepository::findById)
-                .or(() -> categoryRepository.findById(DEFAULT_CATEGORY_ID))
-                .orElseThrow(() -> new GeneralException(ErrorStatus._CATEGORY_NOT_FOUND));
-        return new AiCategoryInfo(category, aiKeywords);
-    }
-    // 3. 감정(Emotion) 조회
-    private Emotion resolveEmotion(Long emotionId) {
-        return (emotionId == null || emotionId <= 0)
-                ? emotionRepository.findById(DEFAULT_EMOTION_ID).orElseThrow(() -> new GeneralException(ErrorStatus._EMOTION_NOT_FOUND))
-                : emotionRepository.findById(emotionId).orElseThrow(() -> new GeneralException(ErrorStatus._EMOTION_NOT_FOUND));
-    }
-    // 4. Domain 조회
-    private Domain resolveDomain(String domainTail) {
-        return domainTail != null
-                ? domainRepository.findByDomainTail(domainTail)
-                .orElseGet(() -> domainRepository.findById(DEFAULT_DOMAIN_ID)
-                        .orElseThrow(() -> new GeneralException(ErrorStatus._DOMAIN_NOT_FOUND)))
-                : domainRepository.findById(DEFAULT_DOMAIN_ID)
-                .orElseThrow(() -> new GeneralException(ErrorStatus._DOMAIN_NOT_FOUND));
-    }
-    // 5. Linku 조회 또는 생성
-    private Linku findOrCreateLinku(String normalizedLink, Category category, Domain domain, String domainTail) {
-        return linkuRepository.findByLinku(normalizedLink)
-                .orElseGet(() -> {
-                    String crawledTitle = linkToImageService.extractTitle(normalizedLink);
-                    if (crawledTitle == null || crawledTitle.isBlank()) {
-                        crawledTitle = (domainTail != null && !domainTail.isBlank()) ? domainTail : "제목 없음";
-                    }
-                    return linkuRepository.save(LinkuConverter.toLinku(normalizedLink, category, domain, crawledTitle));
-                });
-    }
-
-    // 6. AI Article 생성 (기본 키워드 처리 포함, 수정된 부분)
-    private void createAiArticleIfNeeded(Linku linku, Category category, Emotion emotion, String aiKeywords) {
-        if (aiKeywords == null || aiKeywords.isBlank()) {
-            aiKeywords = "키워드 없음"; // 기본키워드로 설정
-        }
-        if (linku.getAiArticle() == null) {
-            Situation defaultSituation = situationRepository.findById(1L)
-                    .orElseThrow(() -> new GeneralException(ErrorStatus._SITUATION_NOT_FOUND));
-            AiArticle aiArticle = AiArticleConverter.toEntityKeywordOnly(aiKeywords, linku, defaultSituation, category, emotion);
-            linku.setAiArticle(aiArticle);
-            aiArticleRepository.save(aiArticle);
-        }
-    }
-    // 7. 사용자 조회
-    private Users findUser(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus._USER_NOT_FOUND));
-    }
-    // 8. 이미지 업로드/추출 처리
-    private String processImage(MultipartFile image, Linku linku) {
-        if (image != null && !image.isEmpty()) {
-            return awsS3Service.uploadFile(image, "linkucreate");
-        }
-        return linkToImageService.getRelatedImageFromUrl(linku.getLinku(), linku.getTitle());
-    }
-    // 9. UsersLinku 생성/저장
-    private UsersLinku createUsersLinku(Users user, Linku linku, Emotion emotion, String memo, String imageUrl) {
-        UsersLinku usersLinku = LinkuConverter.toUsersLinku(user, linku, emotion, memo, imageUrl);
-        return usersLinkuRepository.save(usersLinku);
-    }
-    // 10. 폴더 조회/생성
-    private Folder findOrCreateFolder(Long userId, Category category) {
-        return usersFolderRepository.findFolderByUserIdAndFolderName(userId, category.getCategoryName())
-                .orElseGet(() -> {
-                    Folder newFolder = folderConverter.toFolder(category);
-                    folderRepository.save(newFolder);
-                    UsersFolder newUsersFolder = folderConverter.toUsersFolder(userRepository.getReferenceById(userId), newFolder);
-                    usersFolderRepository.save(newUsersFolder);
-                    return newFolder;
-                });
-    }
-    //링큐 생성 편의 메소드 끝
-
-    @Override
-    @Transactional
     public ApiResponse<LinkuResponseDTO.LinkuIsExistDTO> existLinku(Long userId, String url) {
 
         // 1. 영상 링크 차단 → 예외 던지기
